% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Abuelas_cases.R
\docType{data}
\name{Abuelas_cases}
\alias{Abuelas_cases}
\title{Abuelas cases: Victims of State Terrorism.}
\format{
A data frame with 4 rows and 2 variables:
\describe{
\item{Paternal-Maternal surnames}{Information about the paternal and maternal surnames.}
\item{Birth date}{Birth date of the missing grandchildren}
}
}
\source{
\url{https://www.abuelas.org.ar/caso}
}
\usage{
Abuelas_cases
}
\description{
A dataset containing information from missing grandchildren of Argentina. Four unsolved cases are presented, with the Actual birth date.
}
\keyword{datasets}
