% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/China.R
\docType{data}
\name{China}
\alias{China}
\title{STRs allelic frequencies from specified country.}
\format{
A data frame allele frequencies
}
\usage{
China
}
\description{
STRs allelic frequencies from specified country.
}
\keyword{datasets}
