% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cmodel.R
\name{Cmodel}
\alias{Cmodel}
\title{Epsilon hair color matrix}
\usage{
Cmodel(
  errorModel = c("custom", "uniform")[1],
  ep = 0.01,
  ep12 = 0.01,
  ep13 = 0.005,
  ep14 = 0.01,
  ep15 = 0.003,
  ep23 = 0.01,
  ep24 = 0.003,
  ep25 = 0.01,
  ep34 = 0.003,
  ep35 = 0.003,
  ep45 = 0.01
)
}
\arguments{
\item{errorModel}{custom allows selecting a specfic epsilon for each MP-UHR pair, uniform use ep for all.}

\item{ep}{epsilon}

\item{ep12}{epsilon}

\item{ep13}{epsilon}

\item{ep14}{epsilon}

\item{ep15}{epsilon}

\item{ep23}{epsilon}

\item{ep24}{epsilon}

\item{ep25}{epsilon}

\item{ep34}{epsilon}

\item{ep35}{epsilon}

\item{ep45}{epsilon}
}
\value{
A value of Likelihood ratio based on preliminary investigation data. In this case, sex.
}
\description{
Epsilon hair color matrix
}
\examples{
Cmodel()
}
