% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CondPlot.R
\name{CondPlot}
\alias{CondPlot}
\title{General plot for condiionted probabilities and LR combining variables}
\usage{
CondPlot(CPT_POP, CPT_MP)
}
\arguments{
\item{CPT_POP}{Population conditioned probability table}

\item{CPT_MP}{Missing person conditioned probability table}
}
\value{
A value of Likelihood ratio based on preliminary investigation data. In this case, sex.
}
\description{
General plot for condiionted probabilities and LR combining variables
}
\examples{
Cmodel()
}
