% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPT_MP.R
\name{CPT_MP}
\alias{CPT_MP}
\title{Missing person based conditioned probability}
\usage{
CPT_MP(MPs = "F", MPc = 1, eps = 0.05, epa = 0.05, epc = Cmodel())
}
\arguments{
\item{MPs}{Missing person sex}

\item{MPc}{Missing person hair color}

\item{eps}{sex epsilon}

\item{epa}{age epsilon - Age is not specified in this first version, because it asumes uniformity.}

\item{epc}{color model}
}
\value{
A value of Likelihood ratio based on preliminary investigation data. In this case, sex.
}
\description{
Missing person based conditioned probability
}
\examples{
CPT_MP()
}
