% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CPT_POP.R
\name{CPT_POP}
\alias{CPT_POP}
\title{Population based conditioned probability}
\usage{
CPT_POP(
  propS = c(0.5, 0.5),
  MPa = 40,
  MPr = 6,
  propC = c(0.3, 0.2, 0.25, 0.15, 0.1)
)
}
\arguments{
\item{propS}{age epsilon - Age is not specified in this first version, because it asumes uniformity.}

\item{MPa}{Missing person sex}

\item{MPr}{Missing person hair color}

\item{propC}{sex epsilon}
}
\value{
A value of Likelihood ratio based on preliminary investigation data. In this case, sex.
}
\description{
Population based conditioned probability
}
\examples{
CPT_POP()
}
