% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFreqs.R
\name{getfreqs}
\alias{getfreqs}
\title{Function for getting STR allele frequencies from 48 world populations.}
\usage{
getfreqs(region)
}
\arguments{
\item{region}{select the place of the allele frequency database. Possible values are listed: "Argentina", "Asia","Algeria (M’zab) – Mozabite","Austria","Belgium","Bosnia and Herzegowina","Brazil – Karitiana", "Brazil – Suruí", "Bugainville – NAN Melanesian","Cambodia – Cambodian", "Central African Republic – Biaka Pygmies","China – Dai", "China – Lahu", "China – Yizu", "China – Miaozu", "China – Han", "China – Tu", "China – Oroqen", "China – Hezhen", "China – Uygur", "China – Daur", "China – She", "China – Mongola", "China – Xibo", "China – Naxi", "China – Tujia", "Colombia – Colombian", "Czech Republic", "Democratic Republic of the Congo – Mbuti Pygmies", "Denmark", "Dominican Republic", "Europe", "Finland","France","Germany","Greece","Hungary","Ireland","Israel (Negev) – Bedouin","Israel (Central) – Palestinian","Israel (Carmel) – Druze","Japan – Japanese","Kenya – Bantu (North East)","Mexico – Maya","Mexico – Pima","Montenegro","Namibia – San","New Guinea – Papuan","Nigeria – Yoruba", "Norway", "Pakistan – Burusho", "Pakistan – Sindhi", "Pakistan – Balochi","Pakistan – Kalash","Pakistan – Pathan", "Pakistan – Makrani","Pakistan – Hazara","Poland","Saudi Arabia","Senegal – Mandenka","Siberia - Yakut","Slovenia","Somalia","South Africa – Bantu","Spain","Spain (North-West)","Sweden","Switzerland","Thailand","U.S. All (length-based)","U.S. All (sequence-based)","U.S. African American (length-based)","U.S. African American (sequence-based)","U.S. Asian (sequence-based)", "U.S. Asian (length-based)","U.S. Caucasian (sequence-based)","U.S. Caucasian (length-based)","U.S. Hispanic (length-based)","U.S. Hispanic (sequence-based)"}
}
\value{
An allele frequency database.
}
\description{
Function for getting STR allele frequencies from 48 world populations.
}
\examples{
library(rjson) 
getFreqs("Argentina")
}
