% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRsex.R
\name{LRsex}
\alias{LRsex}
\title{Likelihood ratio for age variable}
\usage{
LRsex(
  MPs = "F",
  eps = 0.05,
  erRs = eps,
  nsims = 1000,
  Ps = c(0.5, 0.5),
  H = 1,
  LR = FALSE
)
}
\arguments{
\item{MPs}{MP sex}

\item{eps}{epsilon paramenter.}

\item{erRs}{error rate in the database.}

\item{nsims}{number of simulations performed.}

\item{Ps}{Sex probabilities in the population.}

\item{H}{hypothesis tested, H1: UHR is MP, H2: UHR is no MP}

\item{LR}{compute LR values}
}
\value{
A value of Likelihood ratio based on preliminary investigation data. In this case, sex.
}
\description{
Likelihood ratio for age variable
}
\examples{
LRsex() 
}
