% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeLRsims.R
\name{makeLRsims}
\alias{makeLRsims}
\title{Make likelihoods ratio (LRs) simulations: a function for obtaining expected LRs under relatedness and unrelatedness kinship hypothesis.}
\usage{
makeLRsims(reference, missing, numsims, verbose = TRUE)
}
\arguments{
\item{reference}{Reference pedigree. It could be an input from read_fam() function or a pedigree built with pedtools.}

\item{missing}{Missing person ID/label indicated in the pedigree.}

\item{numsims}{Number of simulations performed.}

\item{verbose}{Output}
}
\value{
A dataframe with LRs obtained for both hypothesis, Unrelated where POI is not MP or Related where POI is MP.
}
\description{
Make likelihoods ratio (LRs) simulations: a function for obtaining expected LRs under relatedness and unrelatedness kinship hypothesis.
}
\examples{
# A grandchild Search
# library(forrel)
# library(mispitools)
# x = linearPed(2)
# x = setMarkers(x, locusAttributes = NorwegianFrequencies[1:5])
# x = profileSim(x, N = 1, ids = 2)[[1]]
# plot(x)
# makeLRsims(x, missing = 5, 10)

}
