% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{mispitools-deprecated}
\alias{mispitools-deprecated}
\alias{simLRgen}
\alias{simLRprelim}
\alias{simLR2dataframe}
\alias{simRef}
\alias{makePOIprelim}
\alias{makeMPprelim}
\alias{LRsex}
\alias{LRage}
\alias{LRcol}
\alias{LRcolors}
\alias{LRdate}
\alias{combLR}
\alias{compute_LRs_colors}
\alias{CPT_POP}
\alias{CPT_MP}
\alias{Cmodel}
\alias{LRdist}
\alias{deplot}
\alias{CondPlot}
\alias{DeT}
\alias{Trates}
\alias{getfreqs}
\alias{conditionedProp}
\alias{refProp}
\alias{mispiApp}
\alias{lrComparisonApp}
\alias{app_mispitools}
\alias{app_lr_comparison}
\title{Deprecated functions in mispitools}
\usage{
simLRgen(...)

simLRprelim(...)

simLR2dataframe(...)

simRef(...)

makePOIprelim(...)

makeMPprelim(...)

LRsex(...)

LRage(...)

LRcol(...)

LRcolors(...)

LRdate(...)

combLR(...)

compute_LRs_colors(...)

CPT_POP(...)

CPT_MP(...)

Cmodel(...)

LRdist(...)

deplot(...)

CondPlot(...)

DeT(...)

Trates(...)

getfreqs(...)

conditionedProp(...)

refProp(...)

mispiApp(...)

lrComparisonApp(...)

app_mispitools(...)

app_lr_comparison(...)
}
\description{
These functions are deprecated and will be removed in a future version
of mispitools. Please use the new function names instead.

The function names have been updated to follow a consistent snake_case
naming convention with semantic prefixes:
\itemize{
\item \code{sim_*}: Simulation functions
\item \code{lr_*}: Likelihood ratio functions
\item \code{plot_*}: Visualization functions
\item \code{cpt_*}: Conditional probability table functions
\item \code{kl_*}: Kullback-Leibler divergence functions
}
}
\keyword{internal}
