% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marker_model.R
\name{print.marker_model}
\alias{print.marker_model}
\title{Print method for marker_model}
\usage{
\method{print}{marker_model}(x, ...)
}
\arguments{
\item{x}{A \code{marker_model} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for marker_model
}
