% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nongenetic_feature.R
\name{print.nongenetic_feature}
\alias{print.nongenetic_feature}
\title{Print method for nongenetic_feature}
\usage{
\method{print}{nongenetic_feature}(x, ...)
}
\arguments{
\item{x}{A \code{nongenetic_feature} object.}

\item{...}{Unused.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print method for nongenetic_feature
}
