% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/refProp.R
\name{refProp}
\alias{refProp}
\title{Generate a dataframe with hair colour, skin colour, eye colour and their specific combination frequencies}
\usage{
refProp(data)
}
\arguments{
\item{data}{A data.frame containing the characteristics of individuals.}
}
\value{
A data.frame with columns for hair_colour, skin_colour, eye_colour, and f_h_s_y.
}
\description{
This function creates a dataframe that lists every unique combination of hair colour,
skin colour, and eye colour in the provided dataset, along with the proportion of
occurrences of each combination.
}
\examples{
data <- simRef(1000)
refProp(data)
}
