% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLR2dataframe.R
\name{simLR2dataframe}
\alias{simLR2dataframe}
\title{simLR2dataframe: A function for extracting LR distributions in a dataframe from simLRgen() output.}
\usage{
simLR2dataframe(datasim)
}
\arguments{
\item{datasim}{Input dataframe containing expected LRs for related and unrelated POIs. It should be the output from makeLRsims function.}
}
\value{
A dataframe with LR values obtained from simulations.
}
\description{
simLR2dataframe: A function for extracting LR distributions in a dataframe from simLRgen() output.
}
