% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Argentina.R
\docType{data}
\name{Argentina}
\alias{Argentina}
\title{STRs allelic frequencies from specified country.}
\format{
A data frame allele frequencies
}
\usage{
Argentina
}
\description{
STRs allelic frequencies from specified country.
}
\keyword{datasets}
