% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Asia.R
\docType{data}
\name{Asia}
\alias{Asia}
\title{STRs allelic frequencies from specified country.}
\format{
A data frame allele frequencies
}
\usage{
Asia
}
\description{
A dataset of allele frequencies.
}
\keyword{datasets}
