% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Austria.R
\docType{data}
\name{Austria}
\alias{Austria}
\title{STRs allelic frequencies from specified country.}
\format{
A data frame allele frequencies
}
\usage{
Austria
}
\description{
STRs allelic frequencies from specified country.
}
\keyword{datasets}
