% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BosniaHerz.R
\docType{data}
\name{BosniaHerz}
\alias{BosniaHerz}
\title{STRs allelic frequencies from specified country.}
\format{
A data frame allele frequencies
}
\usage{
BosniaHerz
}
\description{
STRs allelic frequencies from specified country.
}
\keyword{datasets}
