% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Europe.R
\docType{data}
\name{Europe}
\alias{Europe}
\title{STRs allelic frequencies from specified country.}
\format{
A data frame allele frequencies
}
\usage{
Europe
}
\description{
STRs allelic frequencies from specified country.
}
\keyword{datasets}
