% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getfreqs.R
\name{getfreqs}
\alias{getfreqs}
\title{Function for getting STR allele frequencies from different world populations.}
\usage{
getfreqs(region)
}
\arguments{
\item{region}{select the place of the allele frequency database. Possible values are listed: "Argentina", "Asia", "Europe", "USA", "Austria", "BosniaHerz", "China" and "Japan". Sources: https://doi.org/10.1016/j.fsigss.2009.08.178; https://doi.org/10.1016/j.fsigen.2016.06.008; https://doi.org/10.1016/j.fsigen.2018.07.013.}
}
\value{
An allele frequency database.
}
\description{
Function for getting STR allele frequencies from different world populations.
}
\examples{
getFfreqs("Argentina")
}
