% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Japan.R
\docType{data}
\name{Japan}
\alias{Japan}
\title{STRs allelic frequencies from specified country.}
\format{
A data frame allele frequencies
}
\usage{
Japan
}
\description{
STRs allelic frequencies from specified country.
}
\keyword{datasets}
