% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRcolors.R
\name{LRcolors}
\alias{LRcolors}
\title{Simulate LR values considering H1 and H2}
\usage{
LRcolors(df, seed = 1234, nsim = 500)
}
\arguments{
\item{df}{A data.frame containing the characteristics of individuals, numerator, f_h_s_y and LRs. Output from compute_LRs function.}

\item{seed}{For replication purposes.}

\item{nsim}{Number of LRs simulated.}
}
\value{
LR distribution considering H1 (Related) and H2 (Unrelated).
}
\description{
Simulate LR values considering H1 and H2
}
