% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mispiApp.R
\name{mispiApp}
\alias{mispiApp}
\title{Missing person shiny app}
\usage{
mispiApp()
}
\value{
An user interface for computing non-genetic LRs and conditioned probability tables.
}
\description{
Missing person shiny app
}
\examples{
CPT_MP()
}
