% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mispitools.R
\docType{package}
\name{mispitools-package}
\alias{mispitools}
\alias{mispitools-package}
\title{mispitools: Missing Person Identification Tools}
\description{
An open source software package written in R statistical language. It consist in a set of decision making tools to conduct missing person searches. Particularly, it allows computing optimal LR threshold for declaring potential matches in DNA-based database search. More recently 'mispitools' incorporates preliminary investigation data based LRs. Statistical weight of different traces of evidence such as biological sex, age and hair color are presented. For citing mispitools please use the following references: Marsico and Caridi, 2023 \doi{10.1016/j.fsigen.2023.102891} and Marsico, Vigeland et al. 2021 \doi{10.1016/j.fsigen.2021.102519}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MarsicoFL/mispitools}
  \item Report bugs at \url{https://github.com/MarsicoFL/mispitools/issues}
}

}
\author{
\strong{Maintainer}: Franco Marsico \email{franco.lmarsico@gmail.com} (\href{https://orcid.org/0000-0002-0740-5516}{ORCID})

}
\keyword{internal}
