% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/USA.R
\docType{data}
\name{USA}
\alias{USA}
\title{STRs allelic frequencies from specified country.}
\format{
A data frame allele frequencies
}
\usage{
USA
}
\description{
STRs allelic frequencies from specified country.
}
\keyword{datasets}
